Reorganize tags interface
